refactor DSL package
