+++ src/test/java/com/amazon/opendistroforelasticsearch/sql/doctest/core/builder/Body.java
@@ -13,7 +13,7 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.doctest.core.dsl;
+package com.amazon.opendistroforelasticsearch.sql.doctest.core.builder;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
